fix rename tests
