+++ src/integration-test/kotlin/integration/EthRelayProviderIrohaTest.kt
@@ -58,7 +58,6 @@ class EthRelayProviderIrohaTest {
      * @when getRelays() is called
      * @then not free wallets are returned in a map
      */
-    @Disabled
     @Test
     fun storageTest() {
         val domain = "notary"
@@ -110,7 +109,6 @@ class EthRelayProviderIrohaTest {
      * @when getRelays() is called
      * @then empty map is returned
      */
-    @Disabled
     @Test
     fun testEmptyStorage() {
         EthRelayProviderIrohaImpl(testConfig.iroha, keypair, detailSetter, detailHolder).getRelays()
