+++ src/integration-test/resources/ethereum_password.properties
@@ -1,4 +1,6 @@
 test.credentialsPassword=user
 withdrawal.credentialsPassword=user
 relay-vacuum.credentialsPassword=user
+ganache.credentialsPassword=user
+
